requirements.txt
pytest~=8.3.4
allure-pytest~=2.13.5
Appium-Python-Client~=4.4.0
selene~=2.0.0rc9
selenium~=4.27.1
allure-python-commons~=2.13.5
pydantic-settings~=2.6.1
python-dotenv~=1.0.1
requests~=2.32.3